requirements.txt
Django==5.1.3
djangorestframework==3.15.2
python-dotenv==0.19.2
asgiref==3.8.1
pytz==2022.1
sqlparse==0.5.1
typing_extensions==4.12.2
coverage==7.6.4